ingest_feeds/jhu_covid19/requirements.txt
gcsfs
pandas
python-dotenv
requests
